rename error file and fix chain-swap network column
