+++ lib/features/transactions/application/application_errors.dart
